Fix table connection: Sample set and individuals tables connected and functional across pages
